@@ -51,46 +51,46 @@
   ],
   "license": "AGPL-3.0",
   "dependencies": {
-    "@babel/runtime": "^7.17.2",
-    "@babel/runtime-corejs3": "^7.17.2",
-    "@tinymce/tinymce-react": "^3.13.1",
+    "@babel/runtime": "^7.17.9",
+    "@babel/runtime-corejs3": "^7.17.9",
+    "@tinymce/tinymce-react": "^4.0.0",
     "bootstrap-sass": "^3.4.3",
     "classnames": "^2.3.1",
     "combine-section-reducers": "^1.0.6",
     "connected-react-router": "^6.9.2",
     "copy-to-clipboard": "^3.3.1",
-    "core-js": "^3.21.1",
+    "core-js": "^3.22.2",
     "deep-equal": "^2.0.5",
     "file-saver": "^2.0.5",
     "history": "<5.0.0",
     "memoize-one": "^6.0.0",
     "param-case": "^3.0.4",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
     "react-dnd-multi-backend": "^6.0.2",
     "react-dnd-touch-backend": "^16.0.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.0.0",
     "react-modal": "^3.14.4",
-    "react-redux": "^7.2.6",
+    "react-redux": "^8.0.1",
     "react-router": "<6.0.0",
     "react-router-dom": "<6.0.0",
     "react-transition-group": "^4.4.2",
     "react-virtualized-auto-sizer": "^1.0.6",
-    "react-window": "^1.8.6",
+    "react-window": "^1.8.7",
     "react-window-infinite-loader": "^1.0.7",
     "reactstrap": "<9.0.0",
-    "redux": "^4.1.2",
+    "redux": "^4.2.0",
     "redux-async-queue": "^1.0.0",
     "redux-thunk": "^2.4.1",
-    "tinymce": "^6.0.0",
+    "tinymce": "^6.0.1",
     "use-debounce": "^7.0.1",
     "zotero-api-client": "^0.39.0",
     "zotero-base-mappings": "0.0.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.5",
+    "@babel/core": "^7.17.9",
     "@babel/eslint-parser": "^7.17.0",
     "@babel/plugin-proposal-class-properties": "^7.16.7",
     "@babel/plugin-transform-async-to-generator": "^7.16.8",
@@ -101,44 +101,44 @@
     "@rollup/plugin-babel": "^5.3.1",
     "@rollup/plugin-commonjs": "^21.0.2",
     "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^13.1.3",
+    "@rollup/plugin-node-resolve": "^13.2.1",
     "@rollup/plugin-replace": "^4.0.0",
     "@rollup/plugin-virtual": "^2.1.0",
-    "@welldone-software/why-did-you-render": "^6.2.3",
-    "autoprefixer": "^10.4.2",
+    "@welldone-software/why-did-you-render": "^7.0.1",
+    "autoprefixer": "^10.4.5",
     "babel-plugin-add-module-exports": "^1.0.4",
     "chai": "^4.3.6",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.6",
-    "eslint": "^8.10.0",
-    "eslint-plugin-react": "^7.29.3",
-    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint": "^8.14.0",
+    "eslint-plugin-react": "^7.29.4",
+    "eslint-plugin-react-hooks": "^4.4.0",
     "fetch-mock": "^9.11.0",
-    "glob": "^7.2.0",
+    "glob": "^8.0.1",
     "http-proxy": "^1.18.1",
-    "mocha": "^9.2.1",
+    "mocha": "^9.2.2",
     "node-fetch": "<3.0.0",
     "nodemon": "^2.0.15",
     "npm-run-all": "^4.1.5",
     "nyc": "^15.1.0",
-    "postcss": "^8.4.8",
+    "postcss": "^8.4.12",
     "postcss-cli": "^9.1.0",
-    "react-test-renderer": "^17.0.2",
+    "react-test-renderer": "^18.0.0",
     "redux-mock-store": "^1.5.4",
     "rimraf": "^3.0.2",
-    "rollup": "^2.70.0",
+    "rollup": "^2.70.2",
     "rollup-plugin-filesize": "^9.1.2",
     "rollup-plugin-sizes": "^1.0.4",
     "rollup-plugin-terser": "^7.0.2",
     "rollup-plugin-web-worker-loader": "^1.6.1",
     "run-sequence": "^2.2.1",
-    "sass": "^1.49.9",
-    "serve-static": "^1.14.2",
-    "sinon": "^13.0.1",
-    "stylelint": "^14.5.3",
-    "stylelint-config-recommended-scss": "^5.0.2",
+    "sass": "^1.51.0",
+    "serve-static": "^1.15.0",
+    "sinon": "^13.0.2",
+    "stylelint": "^14.7.1",
+    "stylelint-config-recommended-scss": "^6.0.0",
     "stylelint-config-standard": "^25.0.0",
     "stylelint-config-standard-scss": "^3.0.0",
-    "stylelint-scss": "^4.1.0"
+    "stylelint-scss": "^4.2.0"
   }
 }
